# server/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "TURISMO ESTG",
        "description": "REST API for web tickets",
        "contact": {
            "email": "8190357@estg.ipp.pt"
        }
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/auth/signUp": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "description": "This can only be done by the logged in user.",
                "operationId": "create User",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/heritage": {
            "get": {
                "tags": [
                    "Heritages"
                ],
                "summary": "Gets all heritages",
                "description": "Gets all heritages from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Heritages List",
                        "schema": {
                            "$ref": "#/definitions/heritage"
                        }
                    }
                },
                "security": [
                    {
                        "item_jwt_auth": []
                    }
                ]
            }
        },
...
        "/promos": {
            "get": {
                "tags": [
                    "Promos"
                ],
                "summary": "Gets all promos",
                "description": "Gets all promos from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Promos List",
                        "schema": {
                            "$ref": "#/definitions/promos"
                        }
                    }
                },
                "security": [
                    {
                        "item_jwt_auth": []
                    }
                ]
            }
        },
...
    },
...
    "definitions": {
        "users": {
            "type": "object",
            "required": [
                "name",
                "email",
                "password"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "SuperMario15",
                    "description": "Name of User"
                },
                "email": {
                    "type": "string",
                    "example": "supermario@gmail.com",
                    "description": "Email of User"
                },
                "password": {
                    "type": "string",
                    "example": "super15mario",
                    "description": "Password of User"
                },
                "roles": {
                    "type": "string",
                    "enum": [
                        "Employee",
                        "User"
                    ]
                },
                "points": {
                    "type": "integer",
                    "format": "int64",
                    "example": "12",
                    "description": "Points of User"
                },
                "createdDate": {
                    "type": "date",
                    "example": "2023-04-08T16:39:05.624+00:00",
                    "description": "Creation Date of User"
                }
            }
        },
        "heritage": {
            "type": "object",
            "required": [
                "heritageImg",
                "name",
                "address",
                "type"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "heritageImg": {
                    "type": "string",
                    "description": "Image of Heritages"
                },
                "name": {
                    "type": "string",
                    "example": "SuperMario15",
                    "description": "Name of Heritages"
                },
                "address": {
                    "type": "string",
                    "example": "Rua das camélias, 45",
                    "description": "Address of Heritages"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "Museu",
                        "Parque de exposição",
                        "Quinta",
                        "Teatro",
                        "Igreja",
                        "Capela",
                        "Aqueduto",
                        "Castelo",
                        "Palacio",
                        "Farol",
                        "Torre",
                        "Estadio",
                        "Ponte",
                        "Zoo",
                        "Oceanário",
                        "Parque Natural"
                    ],
                    "description": "Type of Heritages"
                },
                "visitCounter": {
                    "type": "integer",
                    "format": "int64",
                    "example": "12",
                    "description": "Visit Counter of Heritages"
                }
            }
        },
...
        "promos": {
            "type": "object",
            "required": [
                "namePromo",
                "quantity",
                "percentage"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "namePromo": {
                    "type": "string",
                    "example": "Promo de verão 22/23",
                    "description": "Name of Promo"
                },
                "quantity": {
                    "type": "integer",
                    "format": "int64",
                    "example": "10",
                    "description": "Quantity of Promo"
                },
                "percentage": {
                    "type": "integer",
                    "format": "int64",
                    "example": "15",
                    "description": "Percentage of Promo"
                }
            }
        }
    }
}